+++ tests/structural_map_test.py
@@ -2,10 +2,7 @@
 import pytest
 
 from mets_builder.digital_object import DigitalObject
-from mets_builder.metadata import (DigitalProvenanceAgentMetadata,
-                                   DigitalProvenanceEventMetadata,
-                                   ImportedMetadata, MetadataBase,
-                                   MetadataFormat, MetadataType)
+from mets_builder.metadata import MetadataFormat, MetadataType, MetadataBase
 from mets_builder.structural_map import StructuralMap, StructuralMapDiv
 
 
@@ -105,56 +102,6 @@ def test_add_metadata_to_div():
     )
     div.add_metadata(metadata)
     assert div.metadata == {metadata}
-
-
-def test_add_imported_metadata_to_div():
-    """Test adding imported metadata to a structural map division.
-
-    Metadata import event should be added to div.
-    """
-    div = StructuralMapDiv(div_type="test_type")
-    assert div.metadata == set()
-
-    metadata = ImportedMetadata(
-        metadata_type=MetadataType.DESCRIPTIVE,
-        metadata_format=MetadataFormat.OTHER,
-        other_format="PAS-special",
-        format_version="1.0",
-        data_path="tests/data/imported_metadata.xml"
-    )
-    div.add_metadata(metadata)
-    # In addtition to the added metadata, the div should contain event
-    # metadata
-    assert len(div.metadata) == 2
-    assert metadata in div.metadata
-    event_metadata = (div.metadata - {metadata}).pop()
-    assert event_metadata.event_type == 'metadata extraction'
-    assert event_metadata.event_datetime is None
-    assert event_metadata.event_detail \
-        == "Descriptive metadata import from external source"
-    assert event_metadata.event_outcome.value == "success"
-    assert event_metadata.event_outcome_detail\
-        == "Descriptive metadata imported to mets dmdSec from external source"
-
-
-def test_init_div_with_imported_metadata():
-    """Test initializing div with imported metadata..
-
-    Metadata import event should be added to div.
-    """
-    metadata = ImportedMetadata(
-        metadata_type=MetadataType.DESCRIPTIVE,
-        metadata_format=MetadataFormat.OTHER,
-        other_format="PAS-special",
-        format_version="1.0",
-        data_path="tests/data/imported_metadata.xml"
-    )
-    div = StructuralMapDiv(div_type="test_type", metadata=[metadata])
-
-    assert len(div.metadata) == 2
-    assert metadata in div.metadata
-    event_metadata = (div.metadata - {metadata}).pop()
-    assert event_metadata.event_type == 'metadata extraction'
 
 
 def test_add_digital_objects_to_div():