+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 ----------
 Removed
 ^^^^^^^
+- Creation of automatic PREMIS events when adding imported metadata to structural map. This functionality is moved over to ``siptool-ng``.
 - ``from_directory_structure`` method from ``StructureMap`` class. This functionality is moved over to ``siptool-ng``.
 
 0.3.0 - 2024-03-27